Atualização do Aplication para Jetpack Compose

+++ app/src/main/java/br/com/usinasantafe/pcp/domain/entities/stable/Colab.kt
@@ -1,6 +1,6 @@
 package br.com.usinasantafe.pcp.domain.entities.stable
 
 data class Colab (
-    val matricColab: Long,
+    val matricColab: Int,
     val nomeColab: String,
 )
